Bugfix notification service, added timestamp for substitutes

+++ src/main/java/io/github/paexception/engelsburg/api/controller/reserved/NotificationSettingsController.java
@@ -118,6 +118,6 @@ public class NotificationSettingsController {
 	@Transactional
 	public void deleteInvalidTokens(List<String> invalidTokens) {
 		//Delete all invalid tokens
-		this.tokenRepository.deleteAllByTokenIn(invalidTokens);
+		if (!invalidTokens.isEmpty()) this.tokenRepository.deleteAllByTokenIn(invalidTokens);
 	}
 }
